Modified to maintain fixed ratio of head types released
In the previous version the DQN was learning to produce mostly only one type of part.  To prevent this the factory simulation and the DQN code were modified to make it such that parts are now released in sets subject to a specified part mix.  This is similar to how we were doing it previously with orders.  The part mix specifies the number of each head type that will be released in each order set.  The variable n_part_mix indicates how many order sets will be in process at once.  Once a complete order set is finished, another set is released into the factory.